+Add :? to the variable expansion just to avoid a potential root emptying
Justification: Im always paranoid when running other people's scripts, so i did a shellcheck and umm yeah
https://www.shellcheck.net/wiki/SC2115

+++ amneziawg-install.sh
@@ -926,7 +926,7 @@ function uninstallAmneziaWG() {
 		sysctl --system
 
 		# Remove config files
-		rm -rf ${AMNEZIAWG_DIR}/*
+		rm -rf "${AMNEZIAWG_DIR:?}"/*
 
 		if [[ ${OS} == 'ubuntu' ]]; then
 			apt remove -y amneziawg amneziawg-tools
